refactor(cli): extract context selection test owner
